# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: D8N Core API
  version: 1.0.0
  summary: Multi-brand dating platform API
  description: |
    D8N resolves the active brand from the request host. Never send a brand ID in
    client-controlled input. Bearer sessions are brand-bound, and profile identifiers
    exposed by dating endpoints are public UUIDs rather than internal database IDs.

    ## Implementation status (for frontend integration)

    Every path documented in this contract is implemented and callable. Use this
    summary to see which product domains are complete versus intentionally limited:

    - **Available now:** Identity (register, login, password and login-email
      change, recovery, brand-bound session, identifier verification), Profiles (profile,
      configuration, options, preferences, location, publication), HookUs Matching
      (discovery, likes, passes, matches), DateZA Find (10 unique profiles per
      Johannesburg day) with deterministic DateZA v1 compatibility and
      verification-gated DateZA interactions, and Trust foundations (block,
      report, brand-scoped review and suspension).
    - **Preview (limited scope):** Messaging supports match-gated plain-text
      messages and history. Client idempotency, delivery/read receipts, realtime,
      and message media are not available.
    - **Available (owner-scoped, private):** Profile photos use a direct-to-R2
      upload — request an intent, `PUT` the bytes to the returned short-lived
...

...
    implemented.
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/profiles/{profile_id}/likes:
    post:
      operationId: likeProfile
      tags: [Matching]
      summary: Idempotently like an eligible public profile
      parameters:
        - $ref: "#/components/parameters/PublicProfileId"
      responses:
        "429":
          $ref: "#/components/responses/RateLimited"
        "200":
          description: Existing like returned idempotently.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
        "201":
          description: Like created; a mutual like may also create a match.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/InteractionForbidden"
        "404":
          description: >-
            Matching interactions are not configured for the resolved brand
            (`matching_not_configured`), or the target profile is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          $ref: "#/components/responses/InteractionConflict"
  /api/v1/profiles/{profile_id}/pass:
    post:
      operationId: passProfile
      tags: [Matching]
      summary: Idempotently pass on an eligible public profile
      parameters:
        - $ref: "#/components/parameters/PublicProfileId"
      responses:
        "429":
          $ref: "#/components/responses/RateLimited"
        "200":
          description: Existing pass returned idempotently.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PassResponse"
        "201":
          description: Pass created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PassResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/InteractionForbidden"
        "404":
          description: >-
            Matching interactions are not configured for the resolved brand
            (`matching_not_configured`), or the target profile is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          $ref: "#/components/responses/InteractionConflict"
  /api/v1/profiles/{profile_id}/hook:
    post:
      operationId: sendHook
      tags: [Hooks]
      summary: Send a 🔥 Hook (one opener) to an eligible profile
      description: |
        Creates the sender's single unsolicited opener to a recipient. The body is
        the FINAL text the sender chose — the default copy is a frontend concern
        and receiving it means nothing special. No Match/Conversation/Message is
        created yet: the opener is held on the Hook, and the sender is locked out
        from any further message until the recipient replies (there is literally no
        conversation to message through, and a second Hook to the same person is
        rejected).

        Reachability matches Like exactly: blocked (either direction), suspended,
        closed, hidden, cross-brand, self, and otherwise-ineligible targets all
        fail closed as `profile_unavailable` (404), never revealing which.

        V1 anti-spam: exactly one Hook per (sender → recipient) pair, ever — a
        second attempt (including after decline/expiry) returns `already_hooked`
        (409). Sending is also bounded by a per-sender daily allowance
        (`hook_rate_limited`, 429, with `Retry-After`).
      parameters:
        - $ref: "#/components/parameters/PublicProfileId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendHookRequest"
      responses:
        "201":
          description: The Hook was created and is pending the recipient's reply.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendHookResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: >-
            Hook is not configured for the resolved brand
            (`hook_not_configured`), or the target profile is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: |
            Interaction conflict: `already_hooked` (already used your one Hook to
            this person), `already_matched`, `already_liked`, or `incoming_hook`
            (they Hooked you — reply from your inbox instead).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Blank (`message_blank`) or oversized (`message_too_long`) opener.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: The sender's daily Hook allowance is exhausted (`hook_rate_limited`).
          headers:
            Retry-After:
              description: Seconds until the sender may Hook again.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/hooks/{hook_id}/reply:
    post:
      operationId: replyToHook
      tags: [Hooks]
      summary: Reply to a Hook — accepts it and unlocks a conversation
      description: |
        The recipient's first reply. There is no separate accept step: replying IS
        acceptance. Atomically creates the canonical Match + Conversation,
        materializes the sender's opener as the first message and this text as the
        second, and returns the unlocked conversation. All further messages flow
        through the standard `POST /conversations/{conversation_id}/messages`.

        A Hook that is not live for this member — already accepted/declined,
        expired, blocked either way, or not addressed to the caller — returns the
        neutral `hook_unavailable` (404), never revealing which.
      parameters:
        - $ref: "#/components/parameters/HookId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HookReplyRequest"
      responses:
        "201":
          description: The Hook was accepted; the conversation is unlocked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HookReplyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: >-
            Hook is not configured for the resolved brand
            (`hook_not_configured`), or the Hook is not live for this member
            (`hook_unavailable`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Blank (`message_blank`) or oversized (`message_too_long`) reply.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/hooks/{hook_id}/decline:
    post:
      operationId: declineHook
      tags: [Hooks]
      summary: Decline a pending Hook
      description: |
        The recipient declines. Terminal and private: the sender is never told a
        decline happened (their `hook_state` simply collapses to `unavailable`,
        indistinguishable from being ignored-then-expired or blocked). A Hook that
        is not live for this member returns the neutral `hook_unavailable` (404).
      parameters:
        - $ref: "#/components/parameters/HookId"
      responses:
        "204":
          description: The Hook was declined; nothing is disclosed to the sender.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: >-
            Hook is not configured for the resolved brand
            (`hook_not_configured`), or the Hook is not live for this member
            (`hook_unavailable`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/conversations/{conversation_id}/messages:
    get:
      operationId: listMessages
      tags: [Messaging]
      summary: List a conversation's messages, newest first
      description: |
        Participant-scoped, cursor-paginated plain-text message history for one
        conversation, ordered newest-first by (created_at, id). Only the two
        conversation participants may read. Unknown, cross-brand, non-participant,
        ended, unavailable (suspended/deleted participant), and blocked
        conversations all return the same neutral 404 `conversation_unavailable`;
        the block is never disclosed. History is always bounded by `limit`; page
        backward into older messages with `next_cursor`. A message cursor is bound
        to its brand, viewer, and conversation and cannot be replayed elsewhere.
      parameters:
        - $ref: "#/components/parameters/PublicConversationId"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/MessageLimit"
      responses:
        "200":
          description: A page of messages, newest first, and an opaque next cursor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/InteractionForbidden"
        "404":
          description: >-
            Messaging is not configured for the resolved brand
            (`messaging_not_configured`), or the Conversation is unavailable,
            cross-brand, ended, blocked, or not the current profile's.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/InvalidInput"
    post:
      operationId: sendMessage
      tags: [Messaging]
      summary: Send a plain-text message to a conversation
      description: |
        Appends one plain-text message after re-authorizing the sender
        (participant, active match, both participants available, neither side
        blocking the other). The body is treated as untrusted input, NFC-normalized,
        and must be non-blank and at most 2000 characters; Unicode/emoji is
        preserved. Content never enters logs. Beta accepts at-least-once client
        sends (no idempotency key), so a retried request creates a second message.
      parameters:
        - $ref: "#/components/parameters/PublicConversationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "429":
          $ref: "#/components/responses/RateLimited"
        "201":
          description: The message was persisted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendMessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/InteractionForbidden"
        "404":
          description: >-
            Messaging is not configured for the resolved brand
            (`messaging_not_configured`), or the Conversation is unavailable,
            cross-brand, ended, blocked, or not the current profile's.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Body is blank (`message_blank`) or too long (`message_too_long`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque brand-bound D8N session token.
  parameters:
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque signed cursor returned by the same endpoint and brand.
      schema:
        type: string
...
    PublicProfileId:
      name: profile_id
      in: path
      required: true
      description: Stable public profile UUID from discovery.
      schema:
        type: string
        format: uuid
...
    PublicConversationId:
      name: conversation_id
      in: path
      required: true
      description: Stable public conversation UUID from the conversation list or start-conversation response.
      schema:
        type: string
        format: uuid
    MessageLimit:
      name: limit
      in: query
      required: false
      description: Page size for message history.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 30
    HookId:
      name: hook_id
      in: path
      required: true
      description: Stable public Hook UUID from the received-Hooks inbox.
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Missing, invalid, expired, revoked, or wrong-brand bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InteractionForbidden:
      description: >-
        For a published DateZA member, the identifier attached to the current
        login session has not been verified (`identifier_verification_required`).
        Existing profile, membership, and lifecycle authorization errors remain
        distinct. This DateZA-only rule does not change HookUs authorization.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    RateLimited:
      description: >-
        Rate limit exceeded (`rate_limited`). Returned by authentication throttles
        and by the generic abuse-protection layer on write/read surfaces
        (messaging, likes, passes, reports, media uploads, profile writes,
        discovery, Hook Tonight). The 🔥 Hook send endpoint has its own distinct
        product limit (`hook_rate_limited`) and is documented separately.
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds before another request should be attempted.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InvalidInput:
      description: Request input, cursor, or limit is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InteractionConflict:
      description: Interaction conflicts with the current participant state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ProfileOptionSelections:
      type: object
      additionalProperties:
        type: array
        items:
          type: string
    Language:
      type: object
      required: [code]
      description: |
        A structured language entry. On write, only `code` is required (`proficiency`
        and `primary` optional; at most one primary). On read, `label` is included.
      properties:
        code:
          type: string
          description: Supported taxonomy code (ISO 639-1), e.g. "en".
        label:
          type: string
          readOnly: true
          description: Human-readable label from the taxonomy (read-only).
        proficiency:
          type: [string, "null"]
          enum: [native, fluent, conversational, learning, null]
        primary:
          type: boolean
    ProfileLocation:
      type: object
      required: [precision]
      description: |
        Safe, approximate location metadata for presentation — never raw
        coordinates. Viewer-relative `distance_km` is supplied separately in status.
      properties:
        city:
          type: [string, "null"]
        country_code:
          type: [string, "null"]
        precision:
          type: string
          enum: [approximate]
...
    PublicProfilePhoto:
      type: object
      required: [id, position, url, url_expires_in]
      description: |
        A safe, D8N-generated display derivative of a profile photo. It is
        decoded, re-encoded, and stripped of EXIF/GPS/other metadata. Only
        deliverable photos (visible and finished processing) appear; the raw
        original and its storage key are never exposed. The `url` is a
        short-lived signed URL — refetch the profile to renew it.
      properties:
        id:
          type: string
          format: uuid
          description: |
            Stable, opaque photo id. This is the `target_id` used to report this
            specific photo via POST /api/v1/reports; never the internal id or the
            R2 object key.
        position:
          type: integer
          minimum: 0
        url:
          type: string
          format: uri
        url_expires_in:
          type: integer
    PublicProfile:
      type: object
      description: >-
        Public scalar and derived fields are present only when their source field
        is enabled by the profile's brand. Stable platform envelope fields remain
        present.
      required: [id, photos, options]
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: [string, "null"]
        age:
          type: [integer, "null"]
          minimum: 18
        bio:
          type: [string, "null"]
        gender:
          type: [string, "null"]
        pronouns:
          type: [string, "null"]
        country_code:
          type: [string, "null"]
        city:
          type: [string, "null"]
        location:
          oneOf:
            - $ref: "#/components/schemas/ProfileLocation"
            - type: "null"
        occupation:
          type: [string, "null"]
        job_title:
          type: [string, "null"]
        school_or_institution:
          type: [string, "null"]
        looking_for_text:
          type: [string, "null"]
        height_cm:
          type: [integer, "null"]
        body_type:
          type: [string, "null"]
        languages_spoken:
          type: array
          description: Legacy free-text languages, retained for backward compatibility.
          items:
            type: string
        languages:
          type: array
          description: Canonical structured languages (with taxonomy label).
          items:
            $ref: "#/components/schemas/Language"
        smoking:
          type: [string, "null"]
        drinking:
          type: [string, "null"]
        fitness:
          type: [string, "null"]
        photos:
          type: array
          description: |
            Safe display derivatives, ordered by position. Empty when the profile
            has no photo that is both visible and finished processing.
          items:
            $ref: "#/components/schemas/PublicProfilePhoto"
        options:
          description: |
            Selections for public_profile-visibility groups. On the profile-detail
            endpoint this ALSO includes matches_only groups when an active Match
            exists. company_name and owner-only groups are never included.
          $ref: "#/components/schemas/ProfileOptionSelections"
...
    LikeResponse:
      type: object
      required: [liked, matched, match_id, created]
      properties:
        liked:
          const: true
        matched:
          type: boolean
        match_id:
          type: [string, "null"]
          format: uuid
        created:
          type: boolean
    PassResponse:
      type: object
      required: [passed, created]
      properties:
        passed:
          const: true
        created:
          type: boolean
...
    Conversation:
      type: object
      required: [id, match_id, status, created_at, profile, last_message]
      properties:
        id:
          type: string
          format: uuid
        match_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, closed]
        created_at:
          type: string
          format: date-time
        profile:
          $ref: "#/components/schemas/PublicProfile"
        last_message:
          description: Preview of the newest message, or null for an empty conversation.
          oneOf:
            - $ref: "#/components/schemas/MessagePreview"
            - type: "null"
    MessagePreview:
      type: object
      required: [id, sender_id, body, created_at]
      properties:
        id:
          type: string
          format: uuid
        sender_id:
          type: [string, "null"]
          format: uuid
          description: Public profile id of the author.
        body:
          type: string
        created_at:
          type: string
          format: date-time
    Message:
      type: object
      required: [id, conversation_id, sender_id, body, created_at]
      properties:
        id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        sender_id:
          type: string
          format: uuid
          description: |
            Public profile id of the author. The client identifies its own
            messages by comparing against the counterpart profile id on the
            conversation (there are exactly two participants).
        body:
          type: string
        created_at:
          type: string
          format: date-time
    MessageListResponse:
      type: object
      required: [messages, next_cursor]
      properties:
        messages:
          type: array
          description: Newest-first page of messages.
          items:
            $ref: "#/components/schemas/Message"
        next_cursor:
          type: [string, "null"]
    SendMessageRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 2000
          description: Plain-text message. Non-blank, NFC-normalized, Unicode preserved.
    SendMessageResponse:
      type: object
      required: [message]
      properties:
        message:
          $ref: "#/components/schemas/Message"
...
    SendHookRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 2000
          description: |
            The opener the sender chose to submit (the pre-filled default or an
            edited version). Plain text, non-blank, NFC-normalized, Unicode
            preserved. Same constraints as a chat message.
    SendHookResponse:
      type: object
      required: [hook]
      properties:
        hook:
          $ref: "#/components/schemas/HookAcknowledgement"
    HookAcknowledgement:
      type: object
      description: |
        Minimal sender-facing acknowledgement. The sender learns only that their
        opener is pending and when it lapses — never anything the recipient does.
      required: [id, status, created_at, expires_at]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending]
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          description: When the pending Hook expires if the recipient never engages.
...
    HookReplyRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 2000
          description: The recipient's reply text. Same constraints as a chat message.
    HookReplyResponse:
      type: object
      description: The unlocked conversation plus the recipient's stored reply message.
      required: [conversation, message]
      properties:
        conversation:
          $ref: "#/components/schemas/Conversation"
        message:
          $ref: "#/components/schemas/Message"
...